Support contests longer than a month
Who would do that, right? At least that's what I thought 15 years ago when I picked the Java int type. The max ms in int is about 29 days. (In my defence, it might have even been int minutes that long ago, i.e. 60 months) But GNY decided to leave a test contest running for over a month this year to give teams plenty of time to submit. This makes the contest go into negative time and none of the tools work.

The change is really simple... but completely pervasive. If contests can be longs, then so can freeze times, standings, last submission times, etc. May as well be able to pause a contest a month in advance too. Expect a lot of change, and odds are high I'll miss something in the first commit. :)

+++ CDS/src/org/icpc/tools/cds/presentations/PresentationServer.java
@@ -608,7 +608,7 @@ public class PresentationServer {
 
 						// parse time
 						int index = key.indexOf(":");
-						int time = 0;
+						long time = 0;
 						if (index > 0) {
 							String timeStr = key.substring(0, index);
 							if (timeStr.startsWith("+")) {
@@ -727,13 +727,13 @@ public class PresentationServer {
 		executor.scheduleWithFixedDelay(() -> forEachClient(clients, cl -> cl.writePing()), 30L, 45L, TimeUnit.SECONDS);
 	}
 
-	public void scheduleCallback(IContest c, int contestTimeMs) {
+	public void scheduleCallback(IContest c, long contestTimeMs) {
 		// Future: schedule based on next callback
 		/*// find next event
 		int nextEventTime = Integer.MAX_VALUE;
 		for (TimedEvent event : events)
 			nextEventTime = (int) Math.min(event.contestTimeMs, nextEventTime);
-
+		
 		// leave if there are no scheduled events
 		if (nextEventTime == Integer.MAX_VALUE)
 			return;*/
@@ -746,12 +746,12 @@ public class PresentationServer {
 		double contestTime = System.currentTimeMillis() - startTime;
 		long dt = (long) (contestTimeMs / c.getTimeMultiplier() - contestTime);
 
-		final int nextEventTimeMs = contestTimeMs;
+		final long nextEventTimeMs = contestTimeMs;
 		Trace.trace(Trace.INFO, "Callback scheduled in " + (dt / 1000) + " seconds");
 		executor.schedule(() -> onTime(nextEventTimeMs), dt / 1000, TimeUnit.SECONDS);
 	}
 
-	protected void onTime(int contestTimeMs) {
+	protected void onTime(long contestTimeMs) {
 		List<TimedEvent> remove = new ArrayList<>();
 		for (TimedEvent event : events) {
 			if (event.contestTimeMs <= contestTimeMs) {
